housekeeping/code cleanup
