+++ routes/htmlRoutes.js
@@ -1,181 +1,77 @@
+const express = require('express');
+const htmlRouter = express.Router();
 var db = require('../models');
 
-module.exports = function(app) {
-  app.get('/', function(req, res) {
-    db.indexPage
-      .findOne({
-        where: {
-          id: 1
-        }
-      })
-      .then(testimonial => {
-        res.render('index', { user: req.user, testimonial });
-      });
-  });
-
-  app.get('/signUp', (req, res) => {
-    res.render('signUp');
-  });
-  app.get('/signIn', (req, res) => {
-    res.render('signIn');
-  });
-  app.get('/error', (req, res) => {
-    res.render('error', { message: req.flash('loginMessage')[0] });
-  });
-  // FIXME: THIS NEEDS TO GO AWAY OR BE USED AS THE ONLY CONTROLLER ON /questions
-  // second page reder
-  // app.get('/questions', function(req, res) {
-  //   if (!req.user) {
-  //     console.log("No user exists on questions so I'm going to redirect");
-  //     res.redirect('/');
-  //   } else {
-  //     console.log("seems I'm logged in... so go to questions");
-  //     res.render('questions');
-  //   }
-  // });
-
-  // Proper Profile Page Rendering with View Engine
-  app.get('/profile', function(req, res) {
-    if (!req.user) {
-      res.redirect('/');
-    } else {
-      let userData = {
-        user: req.user,
-        major: {},
-        cost: {}
-      };
-
-      const key = req.user.username ? 'username' : 'localUsername';
-      const value = req.user.username || req.user.localUsername;
-
-      if (!req.user.majorChoice) {
-        res.render('profile', { user: req.user });
-      } else {
-        db.user
-          .findOne({
-            where: {
-              [key]: value
-            }
-          })
-          .then(foundUser => {
-            // MAGIC
-            db.major
-              .findOne({
-                where: {
-                  major: foundUser.majorChoice
-                }
-              })
-              .then(majorRes => {
-                //dataValues to truncate only what you need?
-                userData.major = majorRes.dataValues;
-
-                if (!foundUser.cityChoice) {
-                  null;
-                } else {
-                  db.cost
-                    .findOne({
-                      where: {
-                        city: foundUser.cityChoice
-                      }
-                    })
-                    .then(cityRes => {
-                      const cityResults = cityRes.dataValues;
-                      cityResults.cli_plus_rent = (
-                        (parseInt(cityRes.dataValues.cli_plus_rent) / 100) *
-                        57173
-                      ).toFixed();
-
-                      cityResults.cli = (
-                        (parseInt(cityRes.dataValues.cli) / 100) *
-                        57173
-                      ).toFixed();
-                      userData.cost = cityResults;
-
-                      db.note
-                        .findAll({
-                          where: {
-                            username: foundUser.username
-                          }
-                        })
-                        .then(allNotes => {
-                          if (allNotes) {
-                            res.render('profile', {
-                              user: foundUser,
-                              userData,
-                              notes: allNotes
-                            });
-                          } else {
-                            res.render('profile', {
-                              user: foundUser,
-                              userData
-                            });
-                          }
-                        });
-                    });
-                }
-              });
-          });
+htmlRouter.route('/').get((req, res) => {
+  db.indexPage
+    .findOne({
+      where: {
+        id: 1
       }
+    })
+    .then(testimonial => {
+      res.render('index', { user: req.user, testimonial });
+    });
+});
 
-      /* db.major
-        .findOne({
-          where: {
-            major: req.user.majorChoice
-          }
-        })
-        .then(fullMajorRow => {
+htmlRouter.route('/signUp').get((req, res) => {
+  res.render('signUp');
+});
 
-          console.log("\nI'm fullMajorRow");
-          console.log(fullMajorRow.dataValues);
-          userData.major = fullMajorRow.dataValues;
-          console.log("\n I'm userData with .major updated");
-          console.log(userData);
+htmlRouter.route('/signIn').get((req, res) => {
+  res.render('signIn');
+});
 
-          db.cost
-            .findOne({
-              where: {
-                city: req.user.cityChoice
-              }
-            })
-            .then(fullCostRow => {
-              // console.log("\nI'm fullCostRow");
-              // console.log(fullCostRow.dataValues);
-              // cli_including_rent to USD
-              fullCostRow.dataValues.cli_plus_rent = (
-                (parseInt(fullCostRow.dataValues.cli_plus_rent) / 100) *
+htmlRouter.route('/error').get((req, res) => {
+  res.render('error', { message: req.flash('loginMessage')[0] });
+});
+// Populates all the questions
+htmlRouter.route('/questions').get((req, res) => {
+  if (!req.user) {
+    res.redirect('/');
+  } else {
+    let allQuestions = {
+      major: '',
+      cost: ''
+    };
+    db.major
+      .findAll({ attributes: ['major'] })
+      .then(results => {
+        allQuestions.major = results;
+
+        db.cost
+          .findAll({
+            where: {
+              country: 'United States'
+            },
+            attributes: ['city', 'state']
+          })
+          .then(results => {
+            let arryCliRentModify = results.map(entry => {
+              entry.cli_plus_rent = (
+                (parseInt(entry.cli_plus_rent) / 100) *
                 57173
               ).toFixed();
+            });
 
-              // console.log("\nI'm calculated fullCostRow for cli_plus_rent");
-              // console.log(fullCostRow.dataValues);
+            let arryCliModify = results.map(entry => {
+              entry.cli = ((parseInt(entry.cli) / 100) * 57173).toFixed();
+            });
+            allQuestions.cost = results;
+            res.render('questions', {
+              user: req.user,
+              answers: allQuestions
+            });
+          })
+          .catch(err => console.log(err));
+      })
+      .catch(err => console.log(err));
+  }
+});
 
-              // cli to USD
-              fullCostRow.dataValues.cli = (
-                (parseInt(fullCostRow.dataValues.cli) / 100) *
-                57173
-              ).toFixed();
+// Render 404 page for any unmatched routes
+htmlRouter.route('*').get((req, res) => {
+  res.render('404');
+});
 
-              // console.log("\nI'm calculated fullCostRow for cli");
-              // console.log(fullCostRow.dataValues);
-
-              userData.cost = fullCostRow.dataValues;
-              // console.log("\n I'm userData with .cost updated");
-              // console.log(userData);
-              // res.render('profile', userData);
-            })
-            .then(() => {
-              console.log("\nI'm userData:");
-              console.log(userData);
-              res.render('profile', userData);
-            })
-            .catch(err => console.log(err));
-        })
-        .catch(err => console.log(err)); */
-    }
-  });
-
-  // Render 404 page for any unmatched routes
-  app.get('*', function(req, res) {
-    res.render('404');
-  });
-};
+module.exports = htmlRouter;
